requirements.txt
termcolor
colorama
discord.py
aioconsole
keyboard
requests